requirements.txt
gpiozero==1.5.1
pyserial==3.4
RPi.GPIO==0.7.0
Pillow==5.4.1
pprint==0.1